chore(release): resolve all runtime type annotations and finalize >=85% test coverage

+++ src/demandiq/features/engineer.py
@@ -142,18 +142,24 @@ def build_features(
 
     if "rainfall_mm" in out_df.columns:
         out_df["log_rainfall"] = np.log1p(np.maximum(0.0, out_df["rainfall_mm"]))
-        is_rainy_val = out_df["is_rainy"].to_numpy() if "is_rainy" in out_df.columns else 0
-        out_df["rain_intensity"] = out_df["rainfall_mm"].to_numpy() * is_rainy_val
+        is_rainy_val = (
+            out_df["is_rainy"].to_numpy(dtype=float) if "is_rainy" in out_df.columns else 0.0
+        )
+        out_df["rain_intensity"] = out_df["rainfall_mm"].to_numpy(dtype=float) * is_rainy_val
     else:
         out_df["log_rainfall"] = 0.0
         out_df["rain_intensity"] = 0.0
 
     # 6. Interaction terms and ratios (6 features)
-    promo_val = out_df["promo_active"].to_numpy() if "promo_active" in out_df.columns else 0
-    holiday_val = out_df["is_holiday"].to_numpy() if "is_holiday" in out_df.columns else 0
-    festival_val = out_df["festival_flag"].to_numpy() if "festival_flag" in out_df.columns else 0
-    rain_val = out_df["is_rainy"].to_numpy() if "is_rainy" in out_df.columns else 0
-    weekend_val = out_df["is_weekend"].to_numpy()
+    promo_val = (
+        out_df["promo_active"].to_numpy(dtype=int) if "promo_active" in out_df.columns else 0
+    )
+    holiday_val = out_df["is_holiday"].to_numpy(dtype=int) if "is_holiday" in out_df.columns else 0
+    festival_val = (
+        out_df["festival_flag"].to_numpy(dtype=int) if "festival_flag" in out_df.columns else 0
+    )
+    rain_val = out_df["is_rainy"].to_numpy(dtype=int) if "is_rainy" in out_df.columns else 0
+    weekend_val = out_df["is_weekend"].to_numpy(dtype=int)
 
     out_df["promo_x_weekend"] = (promo_val * weekend_val).astype(int)
     out_df["promo_x_rainy"] = (promo_val * rain_val).astype(int)
